refactor(api): centralize Hono app configuration
