+++ apps/api/src/presentation/http/playlist-action-jobs/routes.ts
@@ -4,56 +4,17 @@ import {
   createJobRequestSchema,
   dismissJobsRequestSchema,
 } from "@playlistwizard/playlist-action-job";
-import { Hono } from "hono";
-import type { PlaylistActionServices } from "../../../composition/playlist-actions";
-import type { Env } from "../../../env";
-import type { AuthSession } from "../../../infrastructure/auth/better-auth";
-import { getTrustedOrigins, isAllowedOrigin } from "../../../shared/config";
-import { JOB_PROGRESS_STREAM_CONNECT_REQUEST_URL } from "../../../shared/job-progress-stream-internal-request";
-import { INITIAL_SNAPSHOT_HEADER } from "../../durable-objects/playlist-action-job-progress-stream";
+import { createHonoApp } from "../hono";
 import { requireSession, requireTrustedOriginForMutation } from "../middleware";
+import { jobProgressRoute } from "./progress/routes";
 
-type Variables = {
-  playlistActions: PlaylistActionServices;
-  session: AuthSession;
-};
-
-export const jobsRoute = new Hono<{
-  Bindings: Env;
-  Variables: Variables;
-}>();
+export const jobsRoute = createHonoApp();
 
 // apply middlewares for /jobs/*
 jobsRoute.use("/*", requireTrustedOriginForMutation);
 jobsRoute.use("/*", requireSession);
 
-jobsRoute.get("/progress", async (c) => {
-  const origin = c.req.header("Origin");
-  if (!isAllowedOrigin(origin, getTrustedOrigins(c.env))) {
-    return c.text("Forbidden", 403);
-  }
-
-  if (c.req.header("Upgrade")?.toLowerCase() !== "websocket") {
-    return c.text("Expected WebSocket upgrade", 426);
-  }
-
-  const session = c.get("session");
-  const { getJobProgressSnapshot } = c.get("playlistActions");
-  const initialSnapshot = await getJobProgressSnapshot(session.user.id);
-  const id = c.env.PLAYLIST_ACTION_JOB_PROGRESS_STREAM.idFromName(
-    session.user.id,
-  );
-  const stub = c.env.PLAYLIST_ACTION_JOB_PROGRESS_STREAM.get(id);
-  const headers = new Headers(c.req.raw.headers);
-  headers.set(INITIAL_SNAPSHOT_HEADER, initialSnapshot);
-
-  return stub.fetch(
-    new Request(JOB_PROGRESS_STREAM_CONNECT_REQUEST_URL, {
-      headers,
-      method: "GET",
-    }),
-  );
-});
+jobsRoute.route("/progress", jobProgressRoute);
 
 jobsRoute.post(
   "/create",
