+++ apps/api/vitest.config.mts
@@ -1,7 +1,9 @@
+import tsconfigPaths from "vite-tsconfig-paths";
 import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   root: import.meta.dirname,
+  plugins: [tsconfigPaths()],
   test: {
     environment: "node",
     include: ["src/**/*.test.ts"],
